pyproject.toml
[project]
name = "imagingservices"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "easyocr>=1.7.2",
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "numpy>=2.4.0",
    "pillow>=12.0.0",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/imagingservices"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.10",
]
